Cambios

+++ src/main/java/com/ista/talento_humano/model/primary/Persona.java
@@ -23,7 +23,7 @@ public class Persona implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_persona")
-    private Integer id_persona;
+    private Long id_persona;
 
     @Column(name = "cedula")
     private String cedula;
